requirements.txt
langchain
openai
fastapi
uvicorn
streamlit
pandas
requests
python-dotenv
langchain-openai
